move doc format util into own file
